[refactore] 랭킹화면 root, action 분리
