+++ lib/presentation/rank/pages/ranking_action.dart
@@ -1,0 +1,8 @@
+sealed class RankingAction {
+  const factory RankingAction.fetchRankings(int level) = FetchRankings;
+}
+
+class FetchRankings implements RankingAction {
+  final int level;
+  const FetchRankings(this.level);
+}
