+++ lib/presentation/rank/pages/ranking_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:jogak_jogak/app/di/app_di.dart';
 import 'package:jogak_jogak/core/constants/app_data.dart';
 import 'package:jogak_jogak/presentation/base/pages/tabbar_page.dart';
-import 'package:jogak_jogak/presentation/rank/widgets/ranking_tabview.dart';
+import 'package:jogak_jogak/presentation/rank/pages/ranking_tabbar_view_root.dart';
 
 class RankingPage extends StatelessWidget {
   const RankingPage({super.key});
@@ -16,6 +16,6 @@ class RankingPage extends StatelessWidget {
       AppData.levels.map((e) => '$e X $e').toList();
   static final List<Widget> _chilren =
       AppData.levels
-          .map((e) => RankingTabview(level: e, viewModel: locator()))
+          .map((e) => RankingTabviewRoot(level: e, viewModel: locator()))
           .toList();
 }
